+++ src/Pages/Chats/index.js
@@ -1,0 +1,3 @@
+import { Chats } from './Chats';
+
+export { Chats };;
