+++ src/Pages/Main/Main.jsx
@@ -1,0 +1,3 @@
+export const Main = () => {
+    return <div>Профиль</div>
+}
