+++ src/Pages/Profile/Profile.jsx
@@ -1,0 +1,5 @@
+export const Profile = () => {
+    return (
+        <div>Профиль</div>)
+}
+
